requirements.txt
numpy==2.4.4
pandas==3.0.2
PyYAML==6.0.3
scipy==1.17.1
torch==2.11.0
yfinance==1.3.0
